refactor: move validation utils under validation folder
Refs: #SFB-39
