+++ app/utils/validation/shape-validators/is-exact-value-added.js
@@ -1,4 +1,4 @@
-import { FORM, RDF } from '../rdflib';
+import { FORM, RDF } from '../../rdflib';
 
 export function isExactValueAddedToExactValueConstraint(store, graph) {
   const exactValueConstraint = store.match(
